+++ runtime/protocol/http-client-engines/http-client-engine-crt/jvm/src/aws/smithy/kotlin/runtime/http/engine/crt/CrtHttpEngine.kt
@@ -79,6 +79,10 @@ public class CrtHttpEngine(public override val config: CrtHttpEngineConfig) : Ht
             }
         }
 
+        callContext.job.invokeOnCompletion {
+            stream.close()
+        }
+
         if (request.isChunked) {
             withContext(SdkDispatchers.IO) {
                 stream.sendChunkedBody(request.body)