+++ runtime/runtime-core/jvm/src/aws/smithy/kotlin/runtime/io/SdkBufferJVM.kt
@@ -37,83 +37,83 @@ public actual class SdkBuffer :
         return inner == other.inner
     }
 
-    override fun skip(byteCount: Long): Unit = commonSkip(byteCount)
+    actual override fun skip(byteCount: Long): Unit = commonSkip(byteCount)
 
-    override fun readByte(): Byte = commonReadByte()
+    actual override fun readByte(): Byte = commonReadByte()
 
-    override fun readShort(): Short = commonReadShort()
+    actual override fun readShort(): Short = commonReadShort()
 
-    override fun readShortLe(): Short = commonReadShortLe()
+    actual override fun readShortLe(): Short = commonReadShortLe()
 
-    override fun readLong(): Long = commonReadLong()
+    actual override fun readLong(): Long = commonReadLong()
 
-    override fun readLongLe(): Long = commonReadLongLe()
+    actual override fun readLongLe(): Long = commonReadLongLe()
 
-    override fun readInt(): Int = commonReadInt()
+    actual override fun readInt(): Int = commonReadInt()
 
-    override fun readIntLe(): Int = commonReadIntLe()
+    actual override fun readIntLe(): Int = commonReadIntLe()
 
-    override fun readAll(sink: SdkSink): Long = commonReadAll(sink)
+    actual override fun readAll(sink: SdkSink): Long = commonReadAll(sink)
 
-    override fun read(sink: ByteArray, offset: Int, limit: Int): Int =
+    actual override fun read(sink: ByteArray, offset: Int, limit: Int): Int =
         commonRead(sink, offset, limit)
 
-    override fun read(sink: SdkBuffer, limit: Long): Long =
+    actual override fun read(sink: SdkBuffer, limit: Long): Long =
         commonRead(sink, limit)
 
     override fun read(dst: ByteBuffer): Int = inner.read(dst)
 
-    override fun readByteArray(): ByteArray = commonReadByteArray()
+    actual override fun readByteArray(): ByteArray = commonReadByteArray()
 
-    override fun readByteArray(byteCount: Long): ByteArray = commonReadByteArray(byteCount)
+    actual override fun readByteArray(byteCount: Long): ByteArray = commonReadByteArray(byteCount)
 
-    override fun readUtf8(): String = commonReadUtf8()
+    actual override fun readUtf8(): String = commonReadUtf8()
 
-    override fun readUtf8(byteCount: Long): String = commonReadUtf8(byteCount)
+    actual override fun readUtf8(byteCount: Long): String = commonReadUtf8(byteCount)
 
-    override fun peek(): SdkBufferedSource = commonPeek()
+    actual override fun peek(): SdkBufferedSource = commonPeek()
 
-    override fun exhausted(): Boolean = commonExhausted()
-    override fun request(byteCount: Long): Boolean = commonRequest(byteCount)
+    actual override fun exhausted(): Boolean = commonExhausted()
+    actual override fun request(byteCount: Long): Boolean = commonRequest(byteCount)
 
-    override fun require(byteCount: Long): Unit = commonRequire(byteCount)
+    actual override fun require(byteCount: Long): Unit = commonRequire(byteCount)
 
-    override fun write(source: ByteArray, offset: Int, limit: Int): Unit =
+    actual override fun write(source: ByteArray, offset: Int, limit: Int): Unit =
         commonWrite(source, offset, limit)
 
-    override fun write(source: SdkSource, byteCount: Long): Unit =
+    actual override fun write(source: SdkSource, byteCount: Long): Unit =
         commonWrite(source, byteCount)
 
-    override fun write(source: SdkBuffer, byteCount: Long): Unit =
+    actual override fun write(source: SdkBuffer, byteCount: Long): Unit =
         commonWrite(source, byteCount)
 
     override fun write(src: ByteBuffer): Int = inner.write(src)
 
-    override fun writeAll(source: SdkSource): Long = commonWriteAll(source)
+    actual override fun writeAll(source: SdkSource): Long = commonWriteAll(source)
 
-    override fun writeUtf8(string: String, start: Int, endExclusive: Int): Unit =
+    actual override fun writeUtf8(string: String, start: Int, endExclusive: Int): Unit =
         commonWriteUtf8(string, start, endExclusive)
 
-    override fun writeByte(x: Byte): Unit = commonWriteByte(x)
+    actual override fun writeByte(x: Byte): Unit = commonWriteByte(x)
 
-    override fun writeShort(x: Short): Unit = commonWriteShort(x)
+    actual override fun writeShort(x: Short): Unit = commonWriteShort(x)
 
-    override fun writeShortLe(x: Short): Unit = commonWriteShortLe(x)
+    actual override fun writeShortLe(x: Short): Unit = commonWriteShortLe(x)
 
-    override fun writeInt(x: Int): Unit = commonWriteInt(x)
+    actual override fun writeInt(x: Int): Unit = commonWriteInt(x)
 
-    override fun writeIntLe(x: Int): Unit = commonWriteIntLe(x)
+    actual override fun writeIntLe(x: Int): Unit = commonWriteIntLe(x)
 
-    override fun writeLong(x: Long): Unit = commonWriteLong(x)
+    actual override fun writeLong(x: Long): Unit = commonWriteLong(x)
 
-    override fun writeLongLe(x: Long): Unit = commonWriteLongLe(x)
+    actual override fun writeLongLe(x: Long): Unit = commonWriteLongLe(x)
 
-    override fun flush(): Unit = commonFlush()
+    actual override fun flush(): Unit = commonFlush()
 
-    override fun emit() {
+    actual override fun emit() {
         inner.emit()
     }
-    override fun close(): Unit = commonClose()
+    actual override fun close(): Unit = commonClose()
     override fun isOpen(): Boolean = inner.isOpen
 
     override fun inputStream(): InputStream = inner.inputStream()